Rename photometricinterpreter.fp to .floatingpoint
