+++ src/test/java/org/apache/commons/imaging/formats/tiff/photometricinterpreters/floatingpoint/PaletteEntryForRangeTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.imaging.formats.tiff.photometricinterpreters.fp;
+package org.apache.commons.imaging.formats.tiff.photometricinterpreters.floatingpoint;
 
 import java.awt.Color;
 import org.junit.jupiter.api.Test;
